# src/docs/swagger.yaml
definitions:
  api_handlers.personData:
    properties:
      first_name:
        maxLength: 10
        minLength: 4
        type: string
      last_name:
        maxLength: 20
        minLength: 6
        type: string
      mobile_number:
        maxLength: 11
        minLength: 11
        type: string
    required:
    - first_name
    - last_name
    - mobile_number
    type: object
...
  github_com_aminazadbakht1_golang-clean-web-api_api_dto.CreatePersianYearRequest:
    properties:
      endAt:
        type: string
      persianTitle:
        maxLength: 4
        minLength: 4
        type: string
      startAt:
        type: string
      year:
        type: integer
    type: object
...
  github_com_aminazadbakht1_golang-clean-web-api_api_dto.Filter:
    properties:
      filterType:
        description: text number
        type: string
      from:
        type: string
      to:
        type: string
      type:
        description: contains notContains equals notEqual startsWith lessThan lessThanOrEqual
          greaterThan greaterThanOrEqual inRange endsWith
        type: string
    type: object
...
  ? github_com_aminazadbakht1_golang-clean-web-api_api_dto.PagedList-github_com_aminazadbakht1_golang-clean-web-api_api_dto_PersianYearResponse
  : properties:
      hasNextPage:
        type: boolean
      hasPreviousPage:
        type: boolean
      items:
        items:
          $ref: '#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_dto.PersianYearResponse'
        type: array
      pageNumber:
        type: integer
      totalPages:
        type: integer
      totalRows:
        type: integer
    type: object
...
  github_com_aminazadbakht1_golang-clean-web-api_api_dto.PaginationInputWithFilter:
    properties:
      filter:
        additionalProperties:
          $ref: '#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_dto.Filter'
        type: object
      pageNumber:
        type: integer
      pageSize:
        type: integer
      sort:
        items:
          $ref: '#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_dto.Sort'
        type: array
    type: object
  github_com_aminazadbakht1_golang-clean-web-api_api_dto.PersianYearResponse:
    properties:
      endAt:
        type: string
      id:
        type: integer
      persianTitle:
        type: string
      startAt:
        type: string
      year:
        type: integer
    type: object
...
  github_com_aminazadbakht1_golang-clean-web-api_api_dto.Sort:
    properties:
      colId:
        type: string
      sort:
        type: string
    type: object
...
  github_com_aminazadbakht1_golang-clean-web-api_api_dto.UpdatePersianYearRequest:
    properties:
      endAt:
        type: string
      persianTitle:
        maxLength: 4
        minLength: 4
        type: string
      startAt:
        type: string
      year:
        type: integer
    type: object
...
  github_com_aminazadbakht1_golang-clean-web-api_api_helper.BaseHttpResponse:
    properties:
      error: {}
      result: {}
      resultCode:
        $ref: '#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_helper.ResultCode'
      success:
        type: boolean
      validationErrors:
        items:
          $ref: '#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_validation.ValidationError'
        type: array
    type: object
  github_com_aminazadbakht1_golang-clean-web-api_api_helper.ResultCode:
    enum:
    - 0
    - 40001
    - 40101
    - 40301
    - 40401
    - 42901
    - 42902
    - 50001
    - 50002
    type: integer
    x-enum-varnames:
    - Success
    - ValidationError
    - AuthError
    - ForbiddenError
    - NotFoundError
    - LimiterError
    - OtpLimiterError
    - CustomRecovery
    - InternalError
  github_com_aminazadbakht1_golang-clean-web-api_api_validation.ValidationError:
    properties:
      message:
        type: string
      property:
        type: string
      tag:
        type: string
      value:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /v1/test/binder/body:
    post:
      consumes:
      - application/json
      description: BodyBinder
      parameters:
      - description: person data
        in: body
        name: person
        required: true
        schema:
          $ref: '#/definitions/api_handlers.personData'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_helper.BaseHttpResponse'
            - properties:
                validationErrors:
                  type: object
              type: object
        "400":
          description: Failed
          schema:
            $ref: '#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: BodyBinder
      tags:
      - Test
  /v1/test/binder/uri/{id}/{name}:
    post:
      consumes:
      - application/json
      description: BodyBinder
      parameters:
      - description: user id
        in: path
        name: id
        required: true
        type: integer
      - description: user name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_helper.BaseHttpResponse'
            - properties:
                validationErrors:
                  type: object
              type: object
        "400":
          description: Failed
          schema:
            $ref: '#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: BodyBinder
      tags:
      - Test
  /v1/test/user/{id}:
    get:
      consumes:
      - application/json
      description: UserById
      parameters:
      - description: user id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_helper.BaseHttpResponse'
        "400":
          description: Failed
          schema:
            $ref: '#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_helper.BaseHttpResponse'
      summary: UserById
      tags:
      - Test
...
  /v1/years/:
    post:
      consumes:
      - application/json
      description: Create a PersianYear
      parameters:
      - description: Create a PersianYear
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_dto.CreatePersianYearRequest'
      responses:
        "201":
          description: PersianYear response
          schema:
            allOf:
            - $ref: '#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_helper.BaseHttpResponse'
            - properties:
                result:
                  $ref: '#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_dto.PersianYearResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Create a PersianYear
      tags:
      - PersianYears
  /v1/years/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a PersianYear
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: response
          schema:
            $ref: '#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_helper.BaseHttpResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_helper.BaseHttpResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Delete a PersianYear
      tags:
      - PersianYears
    get:
      consumes:
      - application/json
      description: Get a PersianYear
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: PersianYear response
          schema:
            allOf:
            - $ref: '#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_helper.BaseHttpResponse'
            - properties:
                result:
                  $ref: '#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_dto.PersianYearResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_helper.BaseHttpResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Get a PersianYear
      tags:
      - PersianYears
    put:
      consumes:
      - application/json
      description: Update a PersianYear
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: integer
      - description: Update a PersianYear
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_dto.UpdatePersianYearRequest'
      responses:
        "200":
          description: PersianYear response
          schema:
            allOf:
            - $ref: '#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_helper.BaseHttpResponse'
            - properties:
                result:
                  $ref: '#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_dto.PersianYearResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_helper.BaseHttpResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Update a PersianYear
      tags:
      - PersianYears
  /v1/years/get-by-filter:
    post:
      consumes:
      - application/json
      description: Get PersianYears
      parameters:
      - description: Request
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_dto.PaginationInputWithFilter'
      responses:
        "200":
          description: PersianYear response
          schema:
            allOf:
            - $ref: '#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_helper.BaseHttpResponse'
            - properties:
                result:
                  $ref: '#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_dto.PagedList-github_com_aminazadbakht1_golang-clean-web-api_api_dto_PersianYearResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Get PersianYears
      tags:
      - PersianYears
securityDefinitions:
  AuthBearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"